+++ galaxy_ml/tests/test_model_validations.py
@@ -12,7 +12,7 @@ from sklearn.utils.testing import assert_equal
 from sklearn.utils.testing import assert_array_equal
 from sklearn.utils.testing import ignore_warnings
 
-from model_validations import (train_test_split, OrderedKFold,
+from galaxy_ml.model_validations import (train_test_split, OrderedKFold,
                                RepeatedOrderedKFold)